+++ packages/antd/src/ToolTipAnt.tsx
@@ -1,5 +1,5 @@
 import { Action, Bind, OneOf } from '@moxb/moxb';
-import { Button, Tooltip } from 'antd';
+import { Button, Tooltip, Spin } from 'antd';
 import { ButtonType } from 'antd/lib/button';
 import { AbstractTooltipProps, RenderFunction } from 'antd/lib/tooltip';
 import { observer } from 'mobx-react';
@@ -55,16 +55,18 @@ export class ToolTipButton extends React.Component<BindToolTipAntProps> {
         if (operation!.invisible) {
             return null;
         }
-        const selectedType = operation!.customData ? 'dashed' : 'ghost';
-        if (operation!.disabled) {
+        const { id, disabled, customData } = operation!;
+        const { fire, pending } = operation as Action;
+        const selectedType = customData ? 'dashed' : 'ghost';
+        if (disabled) {
             return (
                 <Button
-                    data-testid={operation!.id}
+                    data-testid={id}
                     type={type || selectedType}
-                    onClick={(operation as Action).fire}
-                    disabled={operation!.disabled}
+                    onClick={fire}
+                    disabled={disabled}
                     htmlType="button"
-                    icon={icon}
+                    icon={pending ? <Spin /> : icon}
                 >
                     {text}
                 </Button>
@@ -73,12 +75,12 @@ export class ToolTipButton extends React.Component<BindToolTipAntProps> {
             return (
                 <ToolTipAnt operation={operation}>
                     <Button
-                        data-testid={operation!.id}
+                        data-testid={id}
                         type={type || selectedType}
-                        onClick={(operation as Action).fire}
-                        disabled={operation!.disabled}
+                        onClick={fire}
+                        disabled={disabled}
                         htmlType="button"
-                        icon={icon}
+                        icon={pending ? <Spin /> : icon}
                     >
                         {text}
                     </Button>