+++ ChangeLog.md
@@ -4,6 +4,9 @@ Follow the principles in [keepachangelog.com](https://keepachangelog.com)!
 
 # v.Next (Current)
 
+### Added
+ - Support indicating the pending state on ToolTipAnt
+
 # [v0.2.0-beta.51](https://github.com/moxb/moxb/releases/tag/v0.2.0-beta.51) (2020-10-20)
 
 ### Changed